requirements.txt
streamlit==1.31.0
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
timm==0.9.12
numpy==1.26.3
Pillow==10.2.0
scikit-learn==1.4.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.1
